Simplified mcrl2/lts/detail/fixed_vector.h: using private inheritance, it can now be guaranteed that the iterators never are invalidated because the methods that would do so are not available to the user.
Improved comments in mcrl2/lts/detail/liblts_bisim_gjkw.h and in
liblts_bisim_gjkw.cpp: they are now in line with the submitted version of the
article, by using ``RfnB'' instead of ``RefB'' (the latter was too similar to
``RedB'') and ``non-bottom'' instead of ``nonbottom''.

Found a few instances where #include "..." should be #include <...> and vice
versa.


git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@14236 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lts/include/mcrl2/lts/detail/check_complexity.h
@@ -9,9 +9,9 @@
 
 /// \file lts/detail/check_complexity.h
 ///
-/// \brief helper macro for time complexity checks during test runs
+/// \brief helper class for time complexity checks during test runs
 ///
-/// \details We use the macro in this group to check whether the overall time
+/// \details We use the class in this group to check whether the overall time
 /// complexity fits in O(m log n).  Although it is difficult to test this in
 /// general because of the constant factor in the definition of the O()
 /// notation, it is often possible to give a (rather tight) upper bound on the
@@ -24,7 +24,6 @@
 
 #include <cstddef>       // for size_t
 #include <cassert>
-#include <map>
 #include <cmath>         // for log2()
 #include "mcrl2/utilities/logger.h"
 
@@ -110,7 +109,8 @@ class check_complexity
             exit(EXIT_FAILURE);
         }
         // assert(ctr.count <= (SIZE_MAX - max.value) / 200);
-        ctr.percentage = (ctr.count * 200 + max.value) / max.value >> 1;
+        ctr.percentage = (unsigned short)
+                              ((ctr.count * 200 + max.value) / max.value >> 1);
     }
 
     static void stats()
